requirements.txt
alembic==1.7.4
appdirs==1.4.4
asgiref==3.3.1
astroid==2.8.2
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
bcrypt==3.2.0
beautifulsoup4==4.10.0
bleach==3.3.0
blinker==1.4
cachelib==0.4.1
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.9
click==7.1.2
colorama==0.4.4
comtypes==1.1.10
coverage==6.1.2
cryptography==36.0.1
distlib==0.3.1
Django==3.1.5
docopt==0.6.2
ecdsa==0.17.0
filelock==3.0.12
Flask==1.1.4
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-DebugToolbar==0.11.0
Flask-DotEnv==0.1.2
Flask-JWT-Extended==4.3.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Seeder==1.2.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
Flask-UUID==0.2
Flask-WTF==0.15.1
google-api-core==1.30.0
google-auth==1.30.2
google-cloud-bigquery==2.20.0
google-cloud-core==1.6.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpcio==1.38.0
gunicorn==20.1.0
h11==0.13.0
idna==2.10
iniconfig==1.1.1
isort==5.9.3
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.1.0
json-logging==1.3.0
keyboard==0.13.5
lazy-object-proxy==1.6.0
ldap3==2.9.1
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
MouseInfo==0.1.3
mysql-connector-python==8.0.25
mysqlclient==2.0.3
numpy==1.21.2
outcome==1.1.0
packaging==20.9
Pillow==8.4.0
pipenv==2020.11.15
pipreqs==0.4.11
platformdirs==2.4.0
pluggy==1.0.0
proto-plus==1.18.1
protobuf==3.17.3
psycopg==3.0.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///E:/GITLAB/PUBLIC/PYTHON/ai-jarvis/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
PyAutoGUI==0.9.53
pycodestyle==2.7.0
pycparser==2.20
PyGetWindow==0.0.9
pyjokes==0.6.0
PyJWT==2.2.0
pylint==2.11.1
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
PyMsgBox==1.0.9
pymssql==2.2.2
PyMySQL==1.0.2
pyodbc==4.0.32
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
PyRect==0.1.4
PyScreeze==0.1.28
pytest==6.2.5
pytest-flask==1.2.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-jose==3.3.0
pyttsx3==2.90
pytweening==1.0.4
pytz==2020.5
pywhatkit==5.3
pywin32==303
regex==2022.1.18
requests==2.25.1
rsa==4.7.2
secure-smtplib==0.1.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.26
sqlparse==0.4.1
toml==0.10.2
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
urllib3==1.26.5
virtualenv==20.2.2
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
webencodings==0.5.1
Werkzeug==1.0.1
wikipedia==1.4.0
wrapt==1.12.1
wsproto==1.0.0
WTForms==2.3.3
yarg==0.1.9
